refactor: integer
