+++ src/lib.rs
@@ -21,14 +21,12 @@
 //! ```
 #![cfg_attr(not(feature = "std"), no_std)]
 
-use core::{
-    cmp::{max, min},
-    ops::{AddAssign, DivAssign, MulAssign, SubAssign},
-};
-use num_traits::{
-    Bounded, CheckedSub, One, Signed, Zero,
-    ops::checked::{CheckedAdd, CheckedMul},
-};
+use num_traits::Signed;
+
+mod builtin;
+mod integer;
+
+pub use integer::{ascii_to_digit, Integer};
 
 /// Parses an integer from a slice.
 ///
@@ -105,7 +103,7 @@ pub trait FromRadix10: Sized {
 
 /// Types implementing this trait can be parsed from a positional numeral system with radix 10.
 /// Acts much like `FromRadix10`, but performs additional checks for overflows.
-pub trait FromRadix10Checked: FromRadix10 {
+pub trait FromRadix10Checked: Sized {
     /// Parses an integer from a slice.
     ///
     /// # Example
@@ -170,8 +168,8 @@ pub trait FromRadix16: Sized {
 }
 
 /// Types implementing this trait can be parsed from a positional numeral system with radix 16.
-/// Acts much like `FromRadix16`, but performs additional checks for overflows.
-pub trait FromRadix16Checked: FromRadix16 {
+/// Acts much like [`FromRadix16`], but performs additional checks for overflows.
+pub trait FromRadix16Checked: Sized {
     /// Parses an integer from a slice.
     ///
     /// # Example
@@ -292,305 +290,6 @@ pub trait MaxNumDigits {
     fn max_num_digits_negative(radix: Self) -> usize;
 }
 
-impl<I> MaxNumDigits for I
-where
-    I: Bounded + Zero + DivAssign + Ord + Copy,
-{
-    /// Returns the maximum number of digits a nonnegative representation of `I` can have depending
-    /// on `radix`.
-    fn max_num_digits(radix: I) -> usize {
-        let mut max = I::max_value();
-        let mut d = 0;
-        while max > I::zero() {
-            d += 1;
-            max /= radix;
-        }
-        d
-    }
-
-    /// Returns the maximum number of digits a negative representation of `I` can have depending
-    /// on `radix`.
-    fn max_num_digits_negative(radix: I) -> usize {
-        let mut min = I::min_value();
-        let mut d = 0;
-        while min < I::zero() {
-            d += 1;
-            min /= radix;
-        }
-        d
-    }
-}
-
-/// Converts an ascii character to digit
-///
-/// # Example
-///
-/// ```
-/// use atoi::ascii_to_digit;
-/// assert_eq!(Some(5), ascii_to_digit(b'5'));
-/// assert_eq!(None, ascii_to_digit::<u32>(b'x'));
-/// ```
-pub fn ascii_to_digit<I>(character: u8) -> Option<I>
-where
-    I: Zero + One,
-{
-    match character {
-        b'0' => Some(nth(0)),
-        b'1' => Some(nth(1)),
-        b'2' => Some(nth(2)),
-        b'3' => Some(nth(3)),
-        b'4' => Some(nth(4)),
-        b'5' => Some(nth(5)),
-        b'6' => Some(nth(6)),
-        b'7' => Some(nth(7)),
-        b'8' => Some(nth(8)),
-        b'9' => Some(nth(9)),
-        _ => None,
-    }
-}
-
-impl<I> FromRadix10 for I
-where
-    I: Zero + One + AddAssign + MulAssign,
-{
-    fn from_radix_10(text: &[u8]) -> (Self, usize) {
-        let mut index = 0;
-        let mut number = I::zero();
-        while index != text.len() {
-            if let Some(digit) = ascii_to_digit(text[index]) {
-                number *= nth(10);
-                number += digit;
-                index += 1;
-            } else {
-                break;
-            }
-        }
-        (number, index)
-    }
-}
-
-impl<I> FromRadix10Signed for I
-where
-    I: Zero + One + AddAssign + SubAssign + MulAssign,
-{
-    fn from_radix_10_signed(text: &[u8]) -> (Self, usize) {
-        let mut index;
-        let mut number = I::zero();
-
-        let (sign, offset) = text
-            .first()
-            .and_then(|&byte| Sign::try_from(byte))
-            .map(|sign| (sign, 1))
-            .unwrap_or((Sign::Plus, 0));
-
-        index = offset;
-
-        // Having two dedicated loops for both the negative and the nonnegative case is rather
-        // verbose, yet performed up to 40% better then a more terse single loop with
-        // `number += digit * signum`.
-
-        match sign {
-            Sign::Plus => {
-                while index != text.len() {
-                    if let Some(digit) = ascii_to_digit::<I>(text[index]) {
-                        number *= nth(10);
-                        number += digit;
-                        index += 1;
-                    } else {
-                        break;
-                    }
-                }
-            }
-            Sign::Minus => {
-                while index != text.len() {
-                    if let Some(digit) = ascii_to_digit::<I>(text[index]) {
-                        number *= nth(10);
-                        number -= digit;
-                        index += 1;
-                    } else {
-                        break;
-                    }
-                }
-            }
-        }
-
-        (number, index)
-    }
-}
-
-impl<I> FromRadix10SignedChecked for I
-where
-    I: Zero
-        + One
-        + AddAssign
-        + MulAssign
-        + SubAssign
-        + CheckedAdd
-        + CheckedSub
-        + CheckedMul
-        + MaxNumDigits,
-{
-    fn from_radix_10_signed_checked(text: &[u8]) -> (Option<Self>, usize) {
-        let mut index;
-        let mut number = I::zero();
-
-        let (sign, offset) = text
-            .first()
-            .and_then(|&byte| Sign::try_from(byte))
-            .map(|sign| (sign, 1))
-            .unwrap_or((Sign::Plus, 0));
-
-        index = offset;
-
-        // Having two dedicated loops for both the negative and the nonnegative case is rather
-        // verbose, yet performed up to 40% better then a more terse single loop with
-        // `number += digit * signum`.
-
-        match sign {
-            Sign::Plus => {
-                let max_safe_digits = max(1, I::max_num_digits(nth(10))) - 1;
-                let max_safe_index = min(text.len(), max_safe_digits + offset);
-                while index != max_safe_index {
-                    if let Some(digit) = ascii_to_digit::<I>(text[index]) {
-                        number *= nth(10);
-                        number += digit;
-                        index += 1;
-                    } else {
-                        break;
-                    }
-                }
-                // We parsed the digits, which do not need checking now lets see the next one:
-                let mut number = Some(number);
-                while index != text.len() {
-                    if let Some(digit) = ascii_to_digit(text[index]) {
-                        number = number.and_then(|n| n.checked_mul(&nth(10)));
-                        number = number.and_then(|n| n.checked_add(&digit));
-                        index += 1;
-                    } else {
-                        break;
-                    }
-                }
-                (number, index)
-            }
-            Sign::Minus => {
-                let max_safe_digits = max(1, I::max_num_digits_negative(nth(10))) - 1;
-                let max_safe_index = min(text.len(), max_safe_digits + offset);
-                while index != max_safe_index {
-                    if let Some(digit) = ascii_to_digit::<I>(text[index]) {
-                        number *= nth(10);
-                        number -= digit;
-                        index += 1;
-                    } else {
-                        break;
-                    }
-                }
-                // We parsed the digits, which do not need checking now lets see the next one:
-                let mut number = Some(number);
-                while index != text.len() {
-                    if let Some(digit) = ascii_to_digit(text[index]) {
-                        number = number.and_then(|n| n.checked_mul(&nth(10)));
-                        number = number.and_then(|n| n.checked_sub(&digit));
-                        index += 1;
-                    } else {
-                        break;
-                    }
-                }
-                (number, index)
-            }
-        }
-    }
-}
-
-impl<I> FromRadix10Checked for I
-where
-    I: Zero + One + FromRadix10 + CheckedMul + CheckedAdd + MaxNumDigits,
-{
-    fn from_radix_10_checked(text: &[u8]) -> (Option<I>, usize) {
-        let max_safe_digits = max(1, I::max_num_digits_negative(nth(10))) - 1;
-        let (number, mut index) = I::from_radix_10(&text[..min(text.len(), max_safe_digits)]);
-        let mut number = Some(number);
-        // We parsed the digits, which do not need checking now lets see the next one:
-        while index != text.len() {
-            if let Some(digit) = ascii_to_digit(text[index]) {
-                number = number.and_then(|n| n.checked_mul(&nth(10)));
-                number = number.and_then(|n| n.checked_add(&digit));
-                index += 1;
-            } else {
-                break;
-            }
-        }
-        (number, index)
-    }
-}
-
-/// Converts an ascii character to digit
-fn ascii_to_hexdigit<I>(character: u8) -> Option<I>
-where
-    I: Zero + One,
-{
-    match character {
-        b'0' => Some(nth(0)),
-        b'1' => Some(nth(1)),
-        b'2' => Some(nth(2)),
-        b'3' => Some(nth(3)),
-        b'4' => Some(nth(4)),
-        b'5' => Some(nth(5)),
-        b'6' => Some(nth(6)),
-        b'7' => Some(nth(7)),
-        b'8' => Some(nth(8)),
-        b'9' => Some(nth(9)),
-        b'a' | b'A' => Some(nth(10)),
-        b'b' | b'B' => Some(nth(11)),
-        b'c' | b'C' => Some(nth(12)),
-        b'd' | b'D' => Some(nth(13)),
-        b'e' | b'E' => Some(nth(14)),
-        b'f' | b'F' => Some(nth(15)),
-        _ => None,
-    }
-}
-
-impl<I> FromRadix16 for I
-where
-    I: Zero + One + AddAssign + MulAssign,
-{
-    fn from_radix_16(text: &[u8]) -> (Self, usize) {
-        let mut index = 0;
-        let mut number = I::zero();
-        while index != text.len() {
-            if let Some(digit) = ascii_to_hexdigit(text[index]) {
-                number *= nth(16);
-                number += digit;
-                index += 1;
-            } else {
-                break;
-            }
-        }
-        (number, index)
-    }
-}
-
-impl<I> FromRadix16Checked for I
-where
-    I: Zero + One + FromRadix16 + CheckedMul + CheckedAdd + MaxNumDigits,
-{
-    fn from_radix_16_checked(text: &[u8]) -> (Option<I>, usize) {
-        let max_safe_digits = max(1, I::max_num_digits_negative(nth(10))) - 1;
-        let (number, mut index) = I::from_radix_16(&text[..min(text.len(), max_safe_digits)]);
-        let mut number = Some(number);
-        // We parsed the digits, which do not need checking now lets see the next one:
-        while index != text.len() {
-            if let Some(digit) = ascii_to_hexdigit(text[index]) {
-                number = number.and_then(|n| n.checked_mul(&nth(16)));
-                number = number.and_then(|n| n.checked_add(&digit));
-                index += 1;
-            } else {
-                break;
-            }
-        }
-        (number, index)
-    }
-}
-
 /// Representation of a numerical sign
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum Sign {
@@ -627,19 +326,6 @@ impl Sign {
             Sign::Minus => -I::one(),
         }
     }
-}
-
-// At least for primitive types this function does not incur runtime costs, since it is only called
-// with constants
-fn nth<I>(n: u8) -> I
-where
-    I: Zero + One,
-{
-    let mut i = I::zero();
-    for _ in 0..n {
-        i = i + I::one();
-    }
-    i
 }
 
 #[cfg(test)]
